feature: fix comments
