+++ mobile/src/pages/Map/ui/Map.tsx
