+++ mobile/src/pages/index.ts
@@ -1,10 +1,10 @@
-export * from './Home/ui';
-export * from './Catalog/ui';
-export * from './Map/ui';
-export * from './Profile/ui';
-export * from './Cart/ui';
-export * from './Auth/ui';
-export * from './Settings/ui';
-export * from './Product/ui';
-export * from './Orders/ui';
-export * from './PaymentMethods/ui';
+export * from './Home/ui/Home';
+export * from './Catalog/ui/Catalog';
+export * from './Map/ui/Map';
+export * from './Profile/ui/Profile';
+export * from './Cart/ui/Cart';
+export * from './Auth/ui/Auth';
+export * from './Settings/ui/Settings';
+export * from './Product/ui/Product';
+export * from './Orders/ui/Orders';
+export * from './PaymentMethods/ui/PaymentMethods';
