+++ mobile/src/shared/icons/index.ts
@@ -1,0 +1,1 @@
+export * from './Icons';
