+++ mobile/src/shared/ui/CategoryItem/CategoryItem.tsx
