+++ mobile/src/shared/ui/index.ts
@@ -1,5 +1,5 @@
-export * from './CategoryItem';
-export * from './Input';
-export * from './ProductItem';
-export * from './Promo';
-export {ShowError} from './ShowError/ShowError';
+export * from './CategoryItem/CategoryItem';
+export * from './Input/Input';
+export * from './ProductItem/ProductItem';
+export * from './Promo/Promo';
+export * from './ShowError/ShowError';
