+++ mobile/src/widgets/Slider/ui/Slider.tsx
